# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Quant Lab Analysis Service API
  version: 1.0.0
  description: |
    A-share quantitative trading platform REST API.

    The Analysis Service acts as the API gateway, coordinating backtest,
    strategy, data, risk, execution, compliance, and AI research endpoints.

    ## Authentication
    Most endpoints are open in dev mode. When `auth.jwt_secret` is configured,
    protected endpoints require a `Bearer` JWT token obtained via `POST /api/auth/login`.

    ## Conventions
    - All dates use `YYYY-MM-DD` unless noted otherwise.
    - Monetary values are in CNY (Chinese Yuan).
    - Symbols follow the `XXXXXX.SH` / `XXXXXX.SZ` convention.
  contact:
    name: Quant Lab
    url: https://github.com/ruoxizhnya/quant-trading
  license:
    name: MIT
servers:
  - url: http://localhost:8085
    description: Local development server
  - url: http://analysis-service:8085
    description: Docker compose internal URL
...
paths:
...
  /api/backtest:
    get:
      tags: [Backtest]
      summary: List backtest jobs
      description: Returns the most recent backtest jobs.
      operationId: listBacktests
      parameters:
        - name: limit
          in: query
          description: Maximum number of jobs to return (1-100, default 20).
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of backtest jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/BacktestJob'
                  total:
                    type: integer
    post:
      tags: [Backtest]
      summary: Run a backtest or create a backtest job
      description: |
        Accepts two request formats:

        1. **Sync (legacy)**: `strategy` + `stock_pool` + `start_date` + `end_date` —
           runs the backtest synchronously and returns the result.
        2. **Async (job)**: `strategy_id` + `universe` + `start_date` + `end_date` —
           creates an async job and returns the job ID immediately.
      operationId: runBacktest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/BacktestRequest'
                - $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '200':
          description: Synchronous backtest result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResponse'
        '202':
          description: Async job created
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/walkforward:
    post:
      tags: [WalkForward]
      summary: Run walk-forward validation
      operationId: runWalkForward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkForwardRequest'
      responses:
        '200':
          description: Walk-forward report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkForwardReport'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [WalkForward]
      summary: List walk-forward reports
      operationId: listWalkForwardReports
      responses:
        '200':
          description: List of walk-forward reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/WalkForwardReport'
                  count:
                    type: integer
...
  /api/strategies:
    get:
      tags: [Strategies]
      summary: List strategies
      description: Returns all strategies, optionally filtered by type or active status.
      operationId: listStrategies
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        '200':
          description: List of strategies
          content:
            application/json:
              schema:
                type: object
                properties:
                  strategies:
                    type: array
                    items:
                      $ref: '#/components/schemas/StrategyConfig'
    post:
      tags: [Strategies]
      summary: Create a strategy
      operationId: createStrategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrategyRequest'
      responses:
        '200':
          description: Strategy created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  strategy_id:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/factor/compute-ic:
    post:
      tags: [Factor]
      summary: Compute factor IC for a date
      operationId: computeFactorIC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [factor, trade_date]
              properties:
                factor:
                  type: string
                trade_date:
                  type: string
                  format: date
                forward_days:
                  type: integer
                  default: 20
      responses:
        '200':
          description: IC computed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  factor:
                    type: string
                  trade_date:
                    type: string
                  forward_days:
                    type: integer
                  ic:
                    type: number
                  p_value:
                    type: number
                  top_ic:
                    type: number
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================================
  # Copilot (AI strategy generation)
  # ============================================================
...
  /api/screen:
    post:
      tags: [Data Proxy]
      summary: Screen stocks (proxied)
      operationId: screenStocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Screen results
          content:
            application/json:
              schema:
                type: object
        '502':
          description: Data service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained via POST /api/auth/login.

  # ============================================================
  # Parameters
  # ============================================================
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ============================================================
  # Schemas
  # ============================================================
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
        status:
          type: string
...
    BacktestRequest:
      type: object
      required: [strategy, start_date, end_date]
      properties:
        strategy:
          type: string
          example: momentum
        stock_pool:
          type: array
          items:
            type: string
          example: ["000001.SZ", "600000.SH"]
        index_code:
          type: string
          example: "000300.SH"
        start_date:
          type: string
          example: "2024-01-01"
        end_date:
          type: string
          example: "2024-06-01"
        initial_capital:
          type: number
          default: 1000000
        risk_free_rate:
          type: number
          default: 0.03
    CreateJobRequest:
      type: object
      required: [strategy_id, universe, start_date, end_date]
      properties:
        strategy_id:
          type: string
        params:
          type: object
          additionalProperties: true
        universe:
          type: string
          example: "000300.SH"
        start_date:
          type: string
        end_date:
          type: string
        initial_capital:
          type: number
        commission_rate:
          type: number
        slippage_rate:
          type: number
    BacktestResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        strategy:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        total_return:
          type: number
        annual_return:
          type: number
        sharpe_ratio:
          type: number
        sortino_ratio:
          type: number
        max_drawdown:
          type: number
        max_drawdown_date:
          type: string
        win_rate:
          type: number
        total_trades:
          type: integer
        win_trades:
          type: integer
        lose_trades:
          type: integer
        avg_holding_days:
          type: number
        calmar_ratio:
          type: number
        stock_pool:
          type: array
          items:
            type: string
        initial_capital:
          type: number
        portfolio_values:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioValue'
        trades:
          type: array
          items:
            $ref: '#/components/schemas/Trade'
    BacktestJob:
      type: object
      properties:
        id:
          type: string
        strategy_id:
          type: string
        universe:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        result:
          type: object
    PortfolioValue:
      type: object
      properties:
        date:
          type: string
          format: date-time
        total_value:
          type: number
        cash:
          type: number
        positions:
          type: number
    Trade:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        direction:
          type: string
          enum: [long, short, close, hold]
        quantity:
          type: number
        price:
          type: number
        commission:
          type: number
        transfer_fee:
          type: number
        stamp_tax:
          type: number
        timestamp:
          type: string
          format: date-time
        pending_qty:
          type: number
        filled_qty:
          type: number
    WalkForwardRequest:
      type: object
      properties:
        strategy:
          type: string
        stock_pool:
          type: array
          items:
            type: string
        start_date:
          type: string
        end_date:
          type: string
        walk_forward_params:
          type: object
          properties:
            train_days:
              type: integer
              default: 250
            test_days:
              type: integer
              default: 60
            step_days:
              type: integer
              default: 60
            min_train_days:
              type: integer
    WalkForwardReport:
      type: object
      properties:
        strategy_id:
          type: string
        universe:
          type: string
        windows:
          type: array
          items:
            type: object
        avg_test_sharpe:
          type: number
        avg_test_return:
          type: number
        avg_test_max_drawdown:
          type: number
        avg_degradation:
          type: number
        pass_rate:
          type: number
        overall_pass:
          type: boolean
...
    StrategyConfig:
      type: object
      properties:
        id:
          type: integer
        strategy_id:
          type: string
        name:
          type: string
        description:
          type: string
        strategy_type:
          type: string
          enum: [momentum, value, quality, composite]
        params:
          type: string
          description: JSON string of parameters schema.
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateStrategyRequest:
      type: object
      required: [strategy_id, name, strategy_type]
      properties:
        strategy_id:
          type: string
        name:
          type: string
        description:
          type: string
        strategy_type:
          type: string
        params:
          type: object
...